requirements.txt
crewai
langchain-ollama
python-dotenv
